+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-26 12:57:12
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-08-28 11:26:54
+// Commit: feat(test): add bi-directional event bus handler for test
 
 class testComponentState {
-  final String commitRef = '3936';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-08-26T12:57:12+05:00');
+  final String commitRef = '3983';
+  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
+  final DateTime timestamp = DateTime.parse('2026-08-28T11:26:54+05:00');
 }