+++ CHANGELOG.md
@@ -3980,3 +3980,4 @@
 - [2026-08-28 10:32] fix(i18n): prevent double tap duplicate action in i18n
 - [2026-08-28 10:47] docs(security): update readme and API contracts for security
 - [2026-08-28 11:22] test(analytics): add unit tests for analytics business logic
+- [2026-08-28 11:26] feat(test): add bi-directional event bus handler for test